학생 학적 조회 화면

+++ src/main/java/com/msa4lms/domain/auth/services/AuthService.java
@@ -79,7 +79,7 @@ public class AuthService {
     }
 
     // logout
-    public void logout(HttpServletResponse response, int id) {
+    public void logout(HttpServletResponse response, long id) {
         User user = userMapper.findByPk(id);
 
         if(user == null) {
